+++ go-sdk/test/go.mod
@@ -1,36 +1,28 @@
-// Copyright 2023 The Perses Authors
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-module dac
+module dac-test
 
 go 1.23.4
 
-replace github.com/perses/perses => ../../../../../../../ // Use current version
+replace github.com/perses/perses => ../../ // Use current version
 
 require (
 	github.com/perses/perses v0.50.1
 	github.com/perses/plugins/prometheus v0.0.0-20250211164300-9fe475c5d750
-	github.com/perses/plugins/timeserieschart v0.0.0-20250211161353-0a08d3cdcbd5
+	github.com/perses/plugins/staticlistvariable v0.0.0-20250211164300-9fe475c5d750
+	github.com/perses/plugins/table v0.0.0-20250211164300-9fe475c5d750
+	github.com/perses/plugins/timeserieschart v0.0.0-20250211164300-9fe475c5d750
+	github.com/stretchr/testify v1.10.0
 )
 
 require (
 	github.com/beorn7/perks v1.0.1 // indirect
 	github.com/cespare/xxhash/v2 v2.3.0 // indirect
+	github.com/davecgh/go-spew v1.1.2-0.20180830191138-d8f796af33cc // indirect
 	github.com/go-jose/go-jose/v4 v4.0.4 // indirect
 	github.com/jpillora/backoff v1.0.0 // indirect
 	github.com/muhlemmer/gu v0.3.1 // indirect
 	github.com/munnerz/goautoneg v0.0.0-20191010083416-a7dc8b61c822 // indirect
 	github.com/mwitkow/go-conntrack v0.0.0-20190716064945-2f068394615f // indirect
+	github.com/pmezard/go-difflib v1.0.1-0.20181226105442-5d4384ee4fb2 // indirect
 	github.com/prometheus/client_golang v1.20.5 // indirect
 	github.com/prometheus/client_model v0.6.1 // indirect
 	github.com/prometheus/common v0.62.0 // indirect
